code correction

+++ Coditech.Project/Common/Coditech.Common.API.Model/Dashboard/Gym/GymDashboardModel.cs
@@ -7,7 +7,7 @@
         public int TotalMemberCount { get; set; }
         public Decimal ToadysTotalCollection { get; set; }
         public Decimal DaywiseTotalCollection { get; set; }
-        public Decimal TotalLeads { get; set; }
+        public int TotalLeads { get; set; }
         public Decimal DigitalPaymentCollection { get; set; }
         public Decimal ManualPaymentCollection { get; set; }
         public List<GymTransactionOverviewModel> TransactionOverviewList { get; set; }
